🛡️ Sentinel: [security improvement] Harden NoSQL injection detection and fix Cloudflare build
- Enhanced NoSQL injection regex patterns to detect bracket notation and JSON-encoded operators.
- Updated detectSuspiciousPatterns to scan both keys and values in query parameters and headers.
- Added comprehensive tests to verify detection of previously missed NoSQL injection vectors.
- Fixed Cloudflare build failure by renaming src/proxy.ts to src/middleware.ts and configuring Edge runtime.
- Documented learnings in .jules/sentinel.md.
